# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface RegisterDeviceRequest {
  deviceId: string;
  deviceName: string;
}

export interface Device {
  deviceId: string;
  deviceName: string;
  lastSeen: number;
}

export interface DeviceList {
  devices: Device[];
}

export type RelayRequestType = typeof RelayRequestType[keyof typeof RelayRequestType];


export const RelayRequestType = {
  LIST: 'LIST',
  GET: 'GET',
} as const;

export interface RelayRequest {
  requestId: string;
  targetDeviceId: string;
  requesterId: string;
  type: RelayRequestType;
  path: string;
}

export interface RelayRequestResult {
  requestId: string;
  ok: boolean;
}

export interface PendingRequests {
  requests: RelayRequest[];
}

export interface RelayResponse {
  requestId: string;
  data: string;
  error?: string;
}

export interface RelayResult {
  ready: boolean;
  data: string;
  error?: string;
}