# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Device {
  deviceId: string;
  deviceName: string;
  lastSeen: number;
}