# lib/api-zod/src/generated/types/deviceList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Device } from './device';

export interface DeviceList {
  devices: Device[];
}