# lib/api-zod/src/generated/types/pendingRequests.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RelayRequest } from './relayRequest';

export interface PendingRequests {
  requests: RelayRequest[];
}