# lib/api-zod/src/generated/types/relayRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RelayRequestType } from './relayRequestType';

export interface RelayRequest {
  requestId: string;
  targetDeviceId: string;
  requesterId: string;
  type: RelayRequestType;
  path: string;
}