# lib/api-zod/src/generated/types/relayResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelayResponse {
  requestId: string;
  data: string;
  error?: string;
}