# lib/api-zod/src/generated/types/relayResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelayResult {
  ready: boolean;
  data: string;
  error?: string;
}